uints -> u8

+++ bits.odin
@@ -3,23 +3,22 @@ package main
 import "core:intrinsics"
 // import "core:simd/x86"
 
-set_bit :: #force_inline proc(bitboard: ^u64, n: i8) {
-	bitboard^ |= 1 << u8(n);
+set_bit :: #force_inline proc(bitboard: ^u64, n: u8) {
+	bitboard^ |= 1 << n;
 }
 
-get_bit :: #force_inline proc(bitboard: ^u64, n: i8) -> u64{
-	return bitboard^ & (1 << u8(n));
+get_bit :: #force_inline proc(bitboard: ^u64, n: u8) -> u64{
+	return bitboard^ & (1 << n);
 }
 
-clear_bit :: #force_inline proc(bitboard: ^u64, n: i8){
-	bitboard^ &= ~(1 << u8(n));
+clear_bit :: #force_inline proc(bitboard: ^u64, n: u8){
+	bitboard^ &= ~(1 << n);
 }
 
-count_bits :: #force_inline proc(bitboard: u64) -> int{
-    return int(intrinsics.count_ones(bitboard))
-    // return int(x86._popcnt64(bitboard))
+count_bits :: #force_inline proc(bitboard: u64) -> u8{
+    return u8(intrinsics.count_ones(bitboard))
 }
 
-ffs :: #force_inline proc(bitboard: u64) -> int{
-    return int(intrinsics.count_trailing_zeros(bitboard))
+ffs :: #force_inline proc(bitboard: u64) -> u8{
+    return u8(intrinsics.count_trailing_zeros(bitboard))
 }
